+++ lib/rspec/mocks/any_instance/stub_chain.rb
@@ -14,7 +14,13 @@ module RSpec
         def create_message_expectation_on(instance)
           proxy = ::RSpec::Mocks.proxy_for(instance)
           expected_from = IGNORED_BACKTRACE_LINE
-          proxy.add_stub(expected_from, *@expectation_args, &@expectation_block)
+          stub = proxy.add_stub(expected_from, *@expectation_args, &@expectation_block)
+
+          if RSpec::Mocks.configuration.yield_receiver_to_any_instance_implementation_blocks?
+            stub.and_yield_receiver_to_implementation
+          end
+
+          stub
         end
 
         def invocation_order