+++ Changelog.md
@@ -24,6 +24,10 @@ Enhancement:
 * Allow the `have_received` matcher to use a block to set further expectations
   on arguments. (Tim Cowlishaw)
 
+Enhancements:
+
+* Yield the receiver to `any_instance` implementation blocks (Sam Phippen).
+
 ### 2.14.1 / 2013-07-07
 [full changelog](http://github.com/rspec/rspec-mocks/compare/v2.14.0...v2.14.1)